src/requirements.txt
asyncpg 
aiohttp
fastapi
uvicorn
SQLAlchemy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
requests
coverage
jinja2
email-validator
alembic
databases[postgresql]
psycopg2-binary
sqlmodel